requirements.txt
psycopg2
drf-base64
django-filter
djangorestframework
django-crispy-forms
